Add files via upload
Added a Restart button on the Wifi Manager screen.  Sometimes when the power goes out the esp32 won't connect to the router right away and it will connect to the router but show the Wifi Manager screen.  Pressing Restart will reboot it and should go to the Freezer Alarm Server page.